refactor: split API request sources
