+++ API/Makefile
@@ -1,7 +1,7 @@
 TARGET         := API.a
 DEBUG_TARGET   := API_debug.a
 
-SRCS := api_request.cpp api_request_tls.cpp api_promise.cpp api_tls_client.cpp
+SRCS := api_request.cpp api_request_tls.cpp api_request_async.cpp api_promise.cpp api_tls_client.cpp
 
 HEADERS := api.hpp promise.hpp tls_client.hpp
